Update rate limit avoidance
This will wait for a delay proportional to the time until the rate resets based on how much of the rate has been used. Basically, if you used 100% of the capacity, then wait 100% of the time until the rate limit rests, if you only used 50% of the capacity, then only wait 50% of the remaining time.

Signed-off-by: Y. Meyer-Norwood <106889957+norwd@users.noreply.github.com>